Update Service Files

+++ src/main/java/com/List/To_Do/Repository/TarefaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TarefaRepository extends JpaRepository<Tarefa,Long> {
+public interface TarefaRepository extends JpaRepository<Tarefa,Integer> {
 }
